Move ServiceNetworkRegistry implementation
Move the service processor's ServiceNetworkRegistry implementations
under the processor module.  This can now be found at
service::processor::registry.
