+++ libsplinter/src/service/processor/registry.rs
